backend stuff (iteration 4)
User stories affected:

An admin or representative can verify an appointment.
A representative can upload a file that documents eligibility requirements for an accreditation.
A representative can remove an uploaded file that is associated with their organization. The representative must remove the eligibility document for an associated accreditation within that organization.
A student can view the attached document in an accreditation.